pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
dependencies = [
    "PySide6==6.8.2.1",
    "pandas==2.2.2",
    "pyqtgraph==0.13.7",
    "PyYAML==6.0.2",
    "h5py==3.13.0",
    'numba==0.59.1',
    "corner==2.2.2",
    "emcee==3.1.6",
    "matplotlib==3.8.4",
    "numpy>=1.22,<1.27",
    "scipy==1.13.0",
    "bottleneck==1.5.0",
    "scikit-learn==1.4.2",
    "statsmodels==0.14.2",
    "autograd==1.6.2",
    "platformdirs==4.2.1",
    "setuptools>=69",
]
name = "CALISTO"
version = "1.0.0"
authors = [{ name = "Alptug Ulugol" }, { name = "Stefanie Pritzl" }]
description = "A Gui Application for the MT Force Calibration"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = "==3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
]

[project.scripts]
mt-calibrate = "CALISTO.main:main"
calisto = "CALISTO.main:main"
